requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.3
Werkzeug==2.3.7
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
